Island: Move Register endpoint to authentication_service
